Use the correct type, and silence a warning.
llvm-svn: 235126

+++ lldb/source/Plugins/Process/gdb-remote/ProcessGDBRemote.cpp
@@ -3677,7 +3677,7 @@ parseRegisters (xmlDocPtr doc, GDBServerRegisterVec & regList) {
 
     xmlNodePtrVec regs;
     xmlExFindChildren( elm, "reg", regs );
-    for ( int i = 0; i < regs.size( ); i++ ) {
+    for ( unsigned long i = 0; i < regs.size( ); i++ ) {
 
         GdbServerRegisterInfo reg;
         reg.clear( );
